requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.5
django-3-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.12.0
django-six==1.0.4
gunicorn==20.1.0
Pillow==8.3.1
pycodestyle==2.7.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
